rest-service/requirements.txt
flask
flask_smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
marshmallow